chore(files): create `deep.ts` file for common logics
- Remove `deep-copy-array.ts` file
- Remove `deep-merge.ts` file
- Add `deep.ts` file for common logics
- Rename function from `merge` to `deepMerge`
